made form into child component
