+++ frontend/src/Login/Login.css
@@ -1,0 +1,5 @@
+.login-wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
